requirements.txt
Django==2.2.8
django-admin-generator==2.2.0
django-admin-interface==0.12.0
django-colorfield==0.1.15
django-filebrowser-no-grappelli==3.8.0
django-filter==2.2.0
django-seed==0.1.9
drf_dynamic_fields
django-tinymce4-lite==1.7.5
django-utils2==2.6.0
djangorestframework==3.11.0
graphene==2.1.8
graphene-django==2.7.1
graphql-core==2.2.1
graphql-relay==2.0.0
Pillow==6.2.1
django-cors-headers
djangorestframework-api-key
requests